+++ assignment-1.02/opal.h
@@ -5,11 +5,11 @@
 
 extern struct player p;
 
-extern uint8_t room_count;
+extern uint16_t room_count;
 extern struct room *rooms;
 
-extern uint8_t stair_up_count;
-extern uint8_t stair_dn_count;
+extern uint16_t stair_up_count;
+extern uint16_t stair_dn_count;
 
 extern struct stair *stairs_up;
 extern struct stair *stairs_dn;
